% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r
\name{Gb_lab}
\alias{Gb_lab}
\alias{Mb_lab}
\alias{Kb_lab}
\title{Number formatters for scales in base pairs}
\usage{
Gb_lab(x)

Mb_lab(x)

Kb_lab(x)
}
\arguments{
\item{x}{The data (in base pairs) to be formatted as Gb, Mb or Kb}
}
\value{
A character vector with scale labels
}
\description{
For use with \code{\link{ggplot2}}
}
\examples{
\dontrun{
ali <- read_paf(system.file("extdata", "fungi.paf", package="pafr"))
doplot(ali) + scale_x_continuous("Genomic position", label=Mb_lab)
}
}
