% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_query}
\alias{highlight_query}
\alias{highlight_target}
\title{Highlight segments of a query or target genome in a dotplot}
\usage{
highlight_query(ali, bed, ordered_by = c("size", "qstart", "provided"),
  ordering = list(), fill = "yellow", colour = "black", alpha = 0.6)

highlight_target(ali, bed, ordered_by = c("size", "qstart", "provided"),
  fill = "yellow", colour = "black", alpha = 0.6)
}
\arguments{
\item{ali}{pafr or tibble containing a genome alignment (as returned by
\code{\link{read_paf}})}

\item{fill}{character fill colour for highlight segment}

\item{colour, }{character outline colour for highlight segment}

\item{alpha, }{character opacity ([0-1]) for highlight segment}

\item{data.frame}{containing a bed file, as returned by
\code{\link{read_bed}}. Should contain three columns, named "chrom", "start" 
 and "end".}

\item{order_by}{How are the query and target sequences be ordered in the
dotplot this highlight is being added to. This argument should match that
used in \link{\code{dotplot}}.}
}
\description{
This plot is intended to be used in conjunction with \code{link{dotplot}}.
Adding \code{higlight_query} or \code{highlight_target} to a dotplot call
(see examples below) will add a rectangular 'highlight' corresponding to a
particular genomic interval in the corresponding genome.
}
