% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.r
\name{plot_synteny}
\alias{plot_synteny}
\title{Plot synteny between a query and target sequence in a paf alignment}
\usage{
plot_synteny(ali, q_chrom, t_chrom, centre = TRUE, RC = FALSE,
  xlab = "Position in query", ylab = "", x_labeller = Mb_lab)
}
\arguments{
\item{ali}{pafr or tibble containing the genome alignment (as returned by
\code{\link{read_paf}})}

\item{q_chrom}{character, name for the query sequence}

\item{t_chrom}{character, name for the target sequence}

\item{centre, }{logical, if TRUE (default) adjust the position of the target 
sequence so it is centred on the query. If not, both sequences start at
position zero}

\item{RC}{logical, if TRUE, use the reverse and complement for the target
sequence}

\item{xlab, }{character, Name for for x-axis}

\item{ylab, }{character, Name for for x-axus}

\item{x_labeller}{function to be used to label x-axis (defaults to}
}
\value{
A ggplot object dsplaying synteny between query and target sequences
}
\description{
Plot synteny between a query and target sequence in a paf alignment
}
