% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.r
\name{theme_coverage_plot}
\alias{theme_coverage_plot}
\title{A minimalistic ggplot2 theme designed for use with genome coverage plots}
\usage{
theme_coverage_plot(facet_labs = TRUE)
}
\arguments{
\item{facet_labs, }{logical. If TRUE (default) label sequences using the facet
labes. If FALSE sequences are labeled directly using \code{\link{geom_text}}}
}
\description{
A minimalistic ggplot2 theme designed for use with genome coverage plots
}
