% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{highlight_query}
\alias{highlight_query}
\alias{highlight_target}
\title{Highlight segments of a query or target genome in a dot plot}
\usage{
highlight_query(
  ali,
  bed,
  ordered_by = c("size", "qstart", "provided"),
  ordering = list(),
  fill = "yellow",
  colour = "black",
  alpha = 0.6
)

highlight_target(
  ali,
  bed,
  ordered_by = c("size", "qstart", "provided"),
  ordering = list(),
  fill = "yellow",
  colour = "black",
  alpha = 0.6
)
}
\arguments{
\item{ali}{pafr or tibble containing a genome alignment (as returned by
\code{\link{read_paf}})}

\item{bed}{\code{data.frame} or \code{tbl_df} containing a bed file, as returned by
\code{\link{read_bed}}. Should contain three columns named 'chrom', 'start' 
 and 'end'}

\item{ordered_by}{How the query and target sequences are ordered in the
dot plot that this highlight is being added to. This argument should match that
used in \code{\link{dotplot}}.}

\item{ordering}{If \code{order_by} is set to TRUE,
this should be a list with two elements that specify the order of query
and then target sequences in the dot plot. This option is ignored if
\code{order_by} is set to other values}

\item{fill}{character  Fill colour for highlight segment}

\item{colour}{character  Outline colour for highlight segment}

\item{alpha}{character  Opacity ([0-1]) for highlight segment}
}
\description{
This plot is intended to be used in conjunction with \code{link{dotplot}}.
Adding \code{higlight_query} or \code{highlight_target} to a dotplot function call
(see examples below) will add a rectangular 'highlight' corresponding to a
particular genomic interval in the corresponding genome.
}
\examples{
ali <- read_paf( system.file("extdata", "fungi.paf", package="pafr") )
cen <- read_bed(system.file("extdata", "Q_centro.bed", package="pafr"))
dotplot(ali) + highlight_query(ali, cen)
interval <- data.frame(chrom="T_chr3", start=2000000, end=3000000)
dotplot(ali, label_seqs=TRUE) + 
   highlight_target(ali, interval)
}
